+++ app/src/main/java/com/example/devsar/cocktailapp/home/OnDrinkClickListener.kt
@@ -1,0 +1,7 @@
+package com.example.devsar.cocktailapp.home
+
+import com.example.devsar.cocktailapp.home.model.Drink
+
+interface OnDrinkClickListener {
+    fun onDrinkClicked(drink: Drink)
+}
