+++ app/src/main/java/com/example/devsar/cocktailapp/home/drinkAdapter.kt
@@ -10,9 +10,8 @@ import com.example.devsar.cocktailapp.R
 import com.example.devsar.cocktailapp.home.DrinkAdapter.ViewHolder
 import com.example.devsar.cocktailapp.home.model.Drink
 import kotlinx.android.synthetic.main.element_drink_list_info.view.*
-import views.CocktailView
 
-class DrinkAdapter(val listener: CocktailView): RecyclerView.Adapter<ViewHolder>() {
+class DrinkAdapter(val listener: OnDrinkClickListener): RecyclerView.Adapter<ViewHolder>() {
 
     val TAG = "DrinkAdapter"
     var drinkList = mutableListOf<Drink>()
@@ -41,7 +40,7 @@ class DrinkAdapter(val listener: CocktailView): RecyclerView.Adapter<ViewHolder>
                     .load(drink.strDrinkThumb)
                     .into(itemView.drinkImage)
             itemView.setOnClickListener {
-                listener.showMoreInfo(drinkList[adapterPosition])
+                listener.onDrinkClicked(drinkList[adapterPosition])
             }
         }
     }
