+++ app/src/main/java/detailCocktail/CocktailDetailsPresentation.kt
@@ -1,11 +1,8 @@
 package detailCocktail
 
 import com.example.devsar.cocktailapp.home.model.DrinkDetail
+import views.BasePresentation
 
-interface CocktailDetailsPresentation {
+interface CocktailDetailsPresentation: BasePresentation {
     fun showDetail(drink: DrinkDetail)
-    fun startLoading()
-    fun endLoading()
-    fun showError(s: String)
-
 }
