+++ app/src/main/java/views/BasePresentation.kt
@@ -1,0 +1,8 @@
+package views
+
+interface BasePresentation {
+    fun showError(error: String)
+    fun showConnectionError()
+    fun endLoading()
+    fun startLoading()
+}
